refactor: refactoring styles files to publish application
