+++ src/styles/components/Table/styles.ts
