+++ src/styles/pages/styles.ts
@@ -1,13 +1,12 @@
 import styled from 'styled-components'
-const Container = styled.div`
+export const Container = styled.div`
   height: 100vh;
   margin: 0.5rem 0;
 `
 
-const LoadingContainer = styled.div`
+export const LoadingContainer = styled.div`
   height: 100vh;
   display: flex;
   align-items: center;
   justify-content: center;
 `
-export default { Container, LoadingContainer }
